requirements.txt
# Local LLM Policy Gap Analysis - Dependencies
# All packages support offline operation after installation

# Local LLM Framework (choose one)
ollama>=0.1.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# PDF Generation
reportlab>=4.0.0

# Security - File Type Validation (MIME type detection)
# For Windows: python-magic-bin includes libmagic DLL
# For Linux/Mac: install system libmagic first, then python-magic
python-magic-bin==0.4.14; platform_system == "Windows"

# Text Processing

# Web Server (LAN hosting)
flask>=3.0.0
